requirements.txt
cx_Freeze
PySide6-essentials
send2trash
appdirs
autopep8
requests
packaging
pyobjc-framework-Cocoa; sys_platform == "darwin"
google-genai
watchdog